requirements.txt
openai>=1.30.0
streamlit>=1.36.0
python-dotenv>=1.0.1
numpy>=1.26.0
pandas>=2.2.0
pypdf>=4.2.0
docx2txt>=0.8
